Placing order works, get price to 5 decimal places

+++ src/broker/BrokerBase.hpp
@@ -13,7 +13,7 @@ class BrokerBase
         virtual ~BrokerBase() = default;
         virtual int connect() = 0;
         virtual int disconnect() = 0;
-        virtual float getLatestPrice(std::string ticker) = 0;
+        virtual double getLatestPrice(std::string ticker) = 0;
         virtual int placeOrder(oms::Order order) = 0;
         virtual oms::Position getLatestPosition(std::string ticker) = 0;
